requirements.txt
openai==0.28
python-dotenv>=0.19.0
